pyproject.toml
[tool.poetry]
name = "chatgpt-with-streamlit"
version = "0.1.0"
description = ""
authors = ["full-stack-ai <fullstackinfo.ai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.3"
streamlit = "^1.30.0"
python-dotenv = "^1.0.1"
numpy = "^1.26.3"
langchain-openai = "^0.0.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
